requirements.txt
pillow
google-api-python-client
google-auth
google-auth-oauthlib
requests
